feat: Add Buffer.debugData
