+++ docs/developer-guide/debugging.md
@@ -63,6 +63,12 @@ Note that luma.gl also injects and parses `glslify`-style `#define SHADER_NAME` 
 Naming shaders directly in the shader code can help identify which 
 shader is involved when debugging shader parsing errors occur.
 
+## Buffer data inspection
+
+`Buffer` objects contain a CPU side copy of the first few bytes of data in the `buffer.debugData` field. This field is refreshed whenever data is written or read from the CPU side, using `buffer.write()`, `buffer.readAsync()` etc and can be inspected in the debugger to inspect the contents of the buffer.
+
+Note that this CPU side data copy can become invalid when buffers are updated on the GPU by compute shaders or transform feedback operations, in which case reading from the buffer would be necessary to refresh the CPU side data.
+
 ## Parameter Validation
 
 luma.gl runs checks on attributes and buffers when they are being set, catching many trivial errors such as setting uniforms to `undefined` or wrong type (scalar vs array etc).
